# middleware/src/swagger.json
{
  "info": {
    "title": "Test API",
    "version": "1.0.0",
    "description": "Test Express API with autogenerated swagger doc"
  },
  "swagger": "2.0",
  "paths": {
    "/basic": {
      "get": {
        "description": "returns hello world",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "response": "any"
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "Get all podcasts.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "response": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Podcast"
              }
            }
          }
        }
      }
    },
    "/search/{input}": {
      "get": {
        "description": "Get podcasts that match search terms.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "input",
            "description": "Name of podcast, creator or tag you wish to find.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "response": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Podcast"
              }
            }
          }
        }
      }
    },
    "/podcast/{id}": {
      "get": {
        "description": "Get a single podcast.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Podcast Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "response": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Podcast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Podcast": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1234"
        },
        "name": {
          "type": "string",
          "example": "CodeNewbie"
        },
        "cast": {
          "type": "array",
          "items": {
            "type": "string",
            "example": [
              "Saron Yitbarek"
            ]
          }
        },
        "tags": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "1234"
            },
            "name": {
              "type": "string",
              "example": "Tech"
            },
            "count": {
              "type": "integer",
              "example": 25
            }
          }
        },
        "network": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "1234"
            },
            "name": {
              "type": "string",
              "example": "Apple Podcasts"
            },
            "url": {
              "type": "string",
              "example": "https://podcasts.apple.com/us/podcast/id919219256"
            }
          }
        },
        "rating": {
          "type": "number",
          "example": 3.2
        },
        "reviews": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "1234"
            },
            "user": {
              "type": "string",
              "example": "David"
            },
            "rating": {
              "type": "number",
              "example": 4
            },
            "text": {
              "type": "string",
              "example": "This is my podcast review."
            }
          }
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}